move filter for suggestions
